TST: Moved test for gmsh tags to new location
